fix: only pasted clipboard models
GrapesJS typically pastes components in clipboard and those in selection. This would not work well with OMNI and causes fatal lagging problems when many components are being pasted at once in succession in a short period of time. Added a fix that only pastes the components in the clipboard and ignore the selected components.